AI_Employee/pyproject.toml
[project]
name = "ai-employee"
version = "0.1.0"
description = "Bronze Tier Personal AI Employee - Autonomous watcher and action processor for Obsidian vault"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=4.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
ai-employee = "AI_Employee.main:main"

[tool.uv]
dev-dependencies = []
